Move 'Connection' tests into test subdirectory
