+++ test/connection/authenticate.js
@@ -1,50 +1,17 @@
 var should = require('should');
-var helper = require('./test-helper');
+var helper = require('../test-helper');
 
-describe('Connection #refreshToken', function(){
-  var org, oauth;
-
+describe('Connection #authenticate', function(){ 
+  var org;
+  
   beforeEach(function(done){
-    oauth = {
-      access_token: '1234567890',
-      instance_url: 'http://instance.salesforce.com/',
-      refresh_token: '1234567890'
-    };
     org = helper.createConnection({});
     done();
   });
 
-  it('returns an error for missing oauth', function(done){
-    org.refreshToken(null, function(err, res){
+  it('should return error with no OAuth', function(done){
+    org.authenticate({}, function(err, res){
       should.exist(err);
-      err.should.match(/invalid oauth/i);
-      done();
-    });
-  });
-
-  it('returns an error for missing instance url', function(done){
-    delete oauth.instance_url;
-    org.refreshToken(oauth, function(err, res){
-      should.exist(err);
-      err.should.match(/invalid oauth/i);
-      done();
-    });
-  });
-
-  it('returns an error for missing access token', function(done){
-    delete oauth.access_token;
-    org.refreshToken(oauth, function(err, res){
-      should.exist(err);
-      err.should.match(/invalid oauth/i);
-      done();
-    });
-  });
-
-  it('returns an error for missing refresh token', function(done){
-    delete oauth.refresh_token;
-    org.refreshToken(oauth, function(err, res){
-      should.exist(err);
-      err.should.match(/refresh token/i);
       done();
     });
   });
@@ -53,25 +20,25 @@ describe('Connection #refreshToken', function(){
 
     it('parses the response body', function(done){
       helper.setFakewebResponse({statusCode: 200, body: '{"responseKey":"responseValue"}'});
-      org.refreshToken(oauth, function(err, res){
+      org.authenticate({code: 'foo'}, function(err, res){
         should.not.exist(err);
         res.responseKey.should.equal("responseValue");
         done();
       });
     });
-
+  
     it('sets oauth to the body when mode is single', function(done){
-      org = helper.createConnection({mode: 'single', oauth: oauth});
+      org = helper.createConnection({mode: 'single'});
       helper.setFakewebResponse({statusCode: 200, body: '{"responseKey":"responseValue"}'});
-      org.refreshToken(oauth, function(err, res){
+      org.authenticate({code: 'foo'}, function(err, res){
         org.oauth.responseKey.should.equal('responseValue');
         done();
       });
     });
-
+  
     it('responds with error when receiving unparsable JSON', function(done){
       helper.setFakewebResponse({statusCode: 200, body: '<html></html>'});
-      org.refreshToken(oauth, function(err, res){
+      org.authenticate({code: 'foo'}, function(err, res){
         err.should.match(/unparsable json/i);
         res.should.equal('<html></html>');
         done();
@@ -80,7 +47,7 @@ describe('Connection #refreshToken', function(){
 
     it('sets a status code when receiving unparsable JSON', function(done){
       helper.setFakewebResponse({statusCode: 200, body: '<html></html>'});
-      org.refreshToken(oauth, function(err, res){
+      org.authenticate({code: 'foo'}, function(err, res){
         err.statusCode.should.equal(200);
         done();
       });
@@ -92,7 +59,7 @@ describe('Connection #refreshToken', function(){
 
     it('returns without body', function(done){
       helper.setFakewebResponse({statusCode: 400, body: '{"error":"errorValue"}'});
-      org.refreshToken(oauth, function(err, res){
+      org.authenticate({code: 'foo'}, function(err, res){
         should.not.exist(res);
         done();
       });
@@ -100,7 +67,7 @@ describe('Connection #refreshToken', function(){
 
     it('parses the response body', function(done){
       helper.setFakewebResponse({statusCode: 400, body: '{"error":"errorValue"}'});
-      org.refreshToken(oauth, function(err, res){
+      org.authenticate({code: 'foo'}, function(err, res){
         err.should.match(/errorValue/);
         done();
       });
@@ -108,7 +75,7 @@ describe('Connection #refreshToken', function(){
 
     it('sets a status code on the error', function(done){
       helper.setFakewebResponse({statusCode: 400, body: '{"error":"errorValue"}'});
-      org.refreshToken(oauth, function(err, res){
+      org.authenticate({code: 'foo'}, function(err, res){
         err.statusCode.should.equal(400);
         done();
       });
@@ -116,7 +83,7 @@ describe('Connection #refreshToken', function(){
 
     it('responds with specific error when receiving unparsable JSON', function(done){
       helper.setFakewebResponse({statusCode: 400, body: '<html></html>'});
-      org.refreshToken(oauth, function(err, res){
+      org.authenticate({code: 'foo'}, function(err, res){
         err.should.match(/unparsable json/i);
         res.should.equal('<html></html>');
         done();
@@ -125,7 +92,7 @@ describe('Connection #refreshToken', function(){
 
     it('sets a status code when receiving unparsable JSON', function(done){
       helper.setFakewebResponse({statusCode: 400, body: '<html></html>'});
-      org.refreshToken(oauth, function(err, res){
+      org.authenticate({code: 'foo'}, function(err, res){
         err.statusCode.should.equal(400);
         done();
       });
@@ -136,7 +103,7 @@ describe('Connection #refreshToken', function(){
   it('returns the original error', function(done){
     org = helper.createConnection({loginUri: 'bad uri'});
     helper.setFakewebResponse({statusCode: 200, body: '{"responseKey":"responseValue"}'});
-    org.refreshToken(oauth, function(err, res){
+    org.authenticate({code: 'foo'}, function(err, res){
       should.exist(err);
       should.not.exist(res);
       done();
@@ -146,11 +113,11 @@ describe('Connection #refreshToken', function(){
   it('uses the test uri when sandbox environment is specified', function(done){
     org = helper.createConnection({testLoginUri: 'http://hello/', environment: 'sandbox'});
     helper.setFakewebResponse({uri: 'http://hello:80/', statusCode: 200, body: '{"responseKey":"responseValue"}'});
-    org.refreshToken(oauth, function(err, res){
+    org.authenticate({code: 'foo'}, function(err, res){
       should.not.exist(err);
       res.responseKey.should.equal( "responseValue" );
       done();
     });
   });
-
+  
 });
